refactor: now using robo v0.11

+++ src/modules/bulkdelete/bulkdelete.ts
@@ -1,5 +1,5 @@
-import { createCommandConfig } from 'robo.js'
-import type { CommandOptions } from 'robo.js'
+import { createCommandConfig } from '@robojs/discordjs'
+import type { ChatInputCommandInteraction, CommandOptions } from '@robojs/discordjs'
 
 export const config = createCommandConfig({
 	description: 'Delete bulk of messages in a channel',
@@ -17,18 +17,22 @@ export const config = createCommandConfig({
 	]
 } as const)
 
-export default async (interaction, options: CommandOptions<typeof config>) => {
+export default async (interaction: ChatInputCommandInteraction, options: CommandOptions<typeof config>) => {
 	try {
-		const channelId = options.channel.match(/\d+/)[0]
-		const messagesNumber = options.number.match(/\d+/)[0]
+		const channelId = options.channel.match(/\d+/)![0]
+		const messagesNumber = options.number.match(/\d+/)![0]
 
-		const channel = interaction.guild.channels.cache.get(channelId)
+		const channel = interaction.guild?.channels.cache.get(channelId)
+		if (!channel || !channel.isTextBased()) {
+			return { content: 'Channel is not a text-based channel', ephemeral: true }
+		}
+
 		return channel
 			.bulkDelete(Number(messagesNumber))
 			.then(() => {
 				channel.send('Messages deleted successfully')
 			})
-			.cache((error) => {
+			.catch((error) => {
 				channel.send(JSON.stringify(error))
 			})
 	} catch (error) {
